Pytest conversion (#12)
 * Converted all unit tests to pytest
 * Fixed many classifiers case in 2 pass reference algorithms
 * Fixed empty accumulator cases
 * Converted command line call to pytest
 * Fixed flake warnings

+++ statmoments/_native.pyx
@@ -433,13 +433,19 @@ class _bivar_sum_base(_AccBase):
       raise ValueError("The moment should be less or equal than indicated in constructor.")
 
     moments = [m - 1 for m in moments]
-    ma, tr_len, cl_len = self.moment, self.trace_len, self._accs.shape[0]
+    ma, min_cnt = self.moment, self.acc_min_count
+    tr_len, cl_len = self.trace_len, self._accs.shape[0]
     # 2D stat should return a separate piece of memory to avoid buffer corruption
     # while normalizing coskeweness and higher moments
     cm, buf = np.empty((2, maxm, tr_len), dtype=np.float64), self._buf
 
     for ii in range(cl_len):
       n0, n1 = self.counts(ii)
+      if min(n0, n1) < min_cnt:
+        cm[:, moments] = 0
+        yield cm[:, moments]
+        continue
+
       acc0, acc1 = self._accs[ii], self._accs[ii, 1:].T
 
       # Convert to double if accumulator is float
@@ -479,11 +485,16 @@ class _bivar_sum_base(_AccBase):
     ma, min_cnt = self.moment, self.acc_min_count
 
     for ii in range(cl_len):
+      n0, n1 = self.counts(ii)
+      acc0, acc1 = self._accs[ii], self._accs[ii, 1:].T
+
+      if min(n0, n1) < min_cnt:
+        retm[:] = 0
+        yield retm
+        continue
+
       # The left and the right degree of the product terms, e.g. xi^1 * xj^3
       for jj, (lm, rm) in enumerate(zip(*moments)):
-        n0, n1 = self.counts(ii)
-        acc0, acc1 = self._accs[ii], self._accs[ii, 1:].T
-
         # Convert to double if accumulator is float
         acc0 = np.asarray(acc0, dtype=np.float64)
         acc1 = np.asarray(acc1, dtype=np.float64)
@@ -558,10 +569,11 @@ class bivar_sum_detrend(_bivar_sum_base):
   def _moments(self, moments, normalize):
     m1idx = [i for i, m in enumerate(moments) if m == 1]
     has_m1 = len(m1idx) != 0
+    dbuf = self._dbuf if len(self._dbuf) > 0 else 0
     for res in super()._moments(moments, normalize):
       if has_m1:
         # Restore offset if mean requested
-        res[:, m1idx] += self._dbuf
+        res[:, m1idx] += dbuf
       yield res
 
 
@@ -718,7 +730,7 @@ class bivar_cntr(_AccBase):
     return self._cls_count[i]
 
   def _moments(self, moments, normalize):
-    maxm = np.max(moments)
+    maxm, min_cnt = np.max(moments), self.acc_min_count
     if self.moment * 2 < maxm:
       raise ValueError("The moment should be less or equal than indicated in constructor.")
 
@@ -730,23 +742,27 @@ class bivar_cntr(_AccBase):
     retm = np.empty((2, len(moments), tr_len), dtype=np.float64)
 
     for ii in range(cl_len):
+      nn = self._cls_count[ii]
+      if min(nn) < min_cnt:
+        retm[:] = 0
+        yield retm
+        continue
+
       for _i in range(2):
         cm = retm[_i]
         for jj, m in enumerate(moments):
           if m == -1:
             cm[jj] = self._accs1d[ii][_i]
           else:
-            nn = self._cls_count[ii][_i]
-            cm[jj] = 1 / nn * self._accs2d[ii][_i, m // 2, :, m // 2 + m % 2, :].diagonal()
+            cm[jj] = 1 / nn[_i] * self._accs2d[ii][_i, m // 2, :, m // 2 + m % 2, :].diagonal()
 
             if m >= 1 and normalize:
-              sd = np.sqrt(1 / nn * self._accs2d[ii][_i, 0, :, 0, :].diagonal())
+              sd = np.sqrt(1 / nn[_i] * self._accs2d[ii][_i, 0, :, 0, :].diagonal())
               cm[jj] /= sd ** (m + 2)
 
       yield retm
 
   def _comoments(self, moments, normalize):
-    #    self._retm[:] = 0  # TODO: Check if needed
     retm = self._retm
     tr_len, min_cnt = self.trace_len, self.acc_min_count
     triuflatten = _triuflatten_gen(tr_len)
@@ -757,9 +773,13 @@ class bivar_cntr(_AccBase):
         if n0 >= min_cnt:
           C = _calc_comoments(1.0 / n0 * accs2d[0], lm, rm, normalize)
           retm[0, jj] = triuflatten(C)
+        else:
+          retm[0, jj] = 0
         if n1 >= min_cnt:
           C = _calc_comoments(1.0 / n1 * accs2d[1], lm, rm, normalize)
           retm[1, jj] = triuflatten(C)
+        else:
+          retm[1, jj] = 0
       yield self._retm[:, :len(moments[0])]
 
 
@@ -816,7 +836,7 @@ class _BivarNpassBase(_AccBase):
   def __init__(self, tr_len, cl_len, **kwargs):
     super().__init__(tr_len, cl_len, **kwargs)
     self.traces = np.empty((1, tr_len))[1:]
-    self.cls    = np.asarray([bytearray(b'2')], dtype=np.uint8)[1:]  # 2 is non existing element
+    self.cls    = np.asarray([bytearray(b'2' * cl_len)], dtype=np.uint8)[1:]  # 2 is non existing element
     self._tmpsq = np.empty((tr_len, tr_len))  # Has to be square for syrk
     self._retm  = np.empty((2, 2, tr_len * (tr_len + 1) // 2))
 
@@ -842,6 +862,7 @@ class _BivarNpassBase(_AccBase):
     return self.total_count - c1, c1
 
   def _moments(self, moments, normalize):
+    min_cnt = self.acc_min_count
     tr_len, cl_len = self.trace_len, self.classifiers_len
 
     # 2D stat should return a separate piece of memory to avoid buffer corruption
@@ -851,9 +872,16 @@ class _BivarNpassBase(_AccBase):
     for ii in range(cl_len):
       cls0 = (self.cls[:, ii] == 0)
       traces_cl = [self.traces[cls0], self.traces[~cls0]]
+      nn = [len(ts) for ts in traces_cl]
+
+      if min(nn) < min_cnt:
+        retm[:, :len(moments)] = 0
+        yield retm[:, :len(moments)]
+        continue
+
       for _i in range(2):
         tr_set = traces_cl[_i]
-        n, cm = len(tr_set), retm[_i]
+        n, cm = nn[_i], retm[_i]
         tr_set_mean = np.mean(tr_set, axis=0)
 
         for jj, m in enumerate(moments):
@@ -886,14 +914,17 @@ class bivar_2pass(_BivarNpassBase):
     return mem_base + mem_ttstd * 8
 
   def _comoments(self, moments, normalize):
-    C = self._tmpsq
-    tr_len, cl_len = self.trace_len, self.cls.shape[1]
+    min_cnt, C = self.acc_min_count, self._tmpsq
+    tr_len, cl_len = self.trace_len, self.classifiers_len
     for ii in range(cl_len):
       for _i in range(2):
         mft = _sort_meanfree(self.traces[self.cls[:, ii] == _i], False)
         m, n = mft.shape
-        triuflatten = _triuflatten_gen(n)
         retm = self._retm[_i]
+        triuflatten = _triuflatten_gen(n)
+        if m < min_cnt:
+          retm[:] = 0
+          continue
 
         for jj, (lm, rm) in enumerate(zip(*moments)):
           if lm == rm:
@@ -906,7 +937,7 @@ class bivar_2pass(_BivarNpassBase):
             sd = np.std(mft, axis=0)
             retm[jj] /= triuflatten(np.outer(sd ** lm, sd ** rm))
 
-    yield self._retm[:, :len(moments[0])]
+      yield self._retm[:, :len(moments[0])]
 
 
 class bivar_txtbk(_BivarNpassBase):
@@ -938,8 +969,8 @@ class bivar_txtbk(_BivarNpassBase):
     return self._tri[:m]
 
   def _comoments(self, moments, normalize):
-    C = self._tmpsq
-    tr_len, cl_len = self.trace_len, self.cls.shape[1]
+    min_cnt, C = self.acc_min_count, self._tmpsq
+    tr_len, cl_len = self.trace_len, self.classifiers_len
     for ii in range(cl_len):
       cl_set = [(self.cls[:, ii] == 0)]
       cl_set.append(~(cl_set[0]))
@@ -948,6 +979,9 @@ class bivar_txtbk(_BivarNpassBase):
         m, n = mft.shape
         retm = self._retm[_i]
         tri  = self._realloc_tri(len(mft))
+        if m < min_cnt:
+          retm[:] = 0
+          continue
 
         for jj, (lm, rm) in enumerate(zip(*moments)):
           tri = _uni2bivar(mft, C, tri, lm, rm) if lm == rm else _uni2bivar_neq(mft, C, tri, lm, rm)
@@ -959,18 +993,7 @@ class bivar_txtbk(_BivarNpassBase):
             sd = np.std(mft, axis=0)
             retm[jj] /= triuflatten(np.outer(sd ** lm, sd ** rm))
 
-    # Outdated code below, but still gives a hint how to ensure the results
-    # # Double-check result with scipy (fix _lt* with C)
-    # from scipy.stats.stats import _ttest_ind_from_stats as _ttest_stats, _unequal_var_ttest_denom as _uneq_denom
-    # axis, ddof = 0, 1 # ddof=1 to get same t-t result with _uneq_denom
-    # m1, m2 = np.mean(_lt0, axis), np.mean(_lt1, axis)
-    # n1, n2 = _lt0.shape[axis], _lt1.shape[axis]
-    # v1, v2 = np.var(_lt0, axis, ddof), np.var(_lt1, axis, ddof)
-    # df, denom = _uneq_denom(v1, n1, v2, n2)
-    # tstat, pv = _ttest_stats(m1, m2, denom, df)
-    # return tstat
-
-    yield self._retm[:, :len(moments[0])]
+      yield self._retm[:, :len(moments[0])]
 
 
 ##################################### VTK #####################################
@@ -1235,8 +1258,9 @@ class univar_sum_detrend(univar_sum):
   def _moments(self, moments, normalize):
     m1idx = [i for i, m in enumerate(moments) if m == 1]
     has_m1 = len(m1idx) != 0
+    dbuf_nonempty = len(self._dbuf) != 0
     for res in super()._moments(moments, normalize):
-      if has_m1:
+      if has_m1 and dbuf_nonempty:
         # Restore offset if mean requested
         res[:, m1idx] += self._dbuf
       yield res
